Added incentive (kill all enemies before proceed)
Added a signal to SceneManager that can emit when all enemies are dead.

This signal is emited using the all_enemies_died() function, which is called within the new EnemyContainer, when all the enemies die.

The changelevelarea2d connects to the signal on spawn, and if there are no enemies, then monitoring will turn on. Monitoring is off by default